+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GKEService.scala
@@ -6,7 +6,6 @@ import cats.mtl.Ask
 import com.google.api.client.googleapis.javanet.GoogleNetHttpTransport
 import com.google.api.client.json.jackson2.JacksonFactory
 import com.google.api.gax.core.FixedCredentialsProvider
-import com.google.api.gax.rpc.FixedTransportChannelProvider
 import com.google.api.services.container.Container
 import com.google.cloud.container.v1.{ClusterManagerClient, ClusterManagerSettings}
 import com.google.common.util.concurrent.ThreadFactoryBuilder
@@ -74,15 +73,16 @@ object GKEService {
       credential <- credentialResource(pathToCredential.toString)
       credentialsProvider = FixedCredentialsProvider.create(credential)
       executorProviderBuilder = ClusterManagerSettings.defaultExecutorProviderBuilder()
-      executorProvider = getExecutorProvider(executorProviderBuilder, "goog2-cluster-manager-%d")
-      transportProvider =
-        ClusterManagerSettings.defaultTransportChannelProvider().withExecutor(executorProvider.getExecutor)
+      threadFactory = new ThreadFactoryBuilder()
+        .setThreadFactory(executorProviderBuilder.getThreadFactory)
+        .setNameFormat("goog2-cluster-manager-%d")
+        .build()
+      executorProvider = executorProviderBuilder.setThreadFactory(threadFactory).build()
 
       clusterManagerSettings = ClusterManagerSettings
         .newBuilder()
         .setCredentialsProvider(credentialsProvider)
         .setBackgroundExecutorProvider(executorProvider)
-        .setTransportChannelProvider(transportProvider)
         .build()
       clusterManager <- backgroundResourceF(ClusterManagerClient.create(clusterManagerSettings))
       legacyClient <- legacyClient(pathToCredential)