+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GoogleTopicAdminInterpreter.scala
@@ -4,7 +4,6 @@ import cats.effect.{Async, Resource}
 import cats.mtl.Ask
 import cats.syntax.all._
 import com.google.api.gax.core.FixedCredentialsProvider
-import com.google.api.gax.rpc.FixedTransportChannelProvider
 import com.google.auth.oauth2.ServiceAccountCredentials
 import com.google.cloud.Identity
 import com.google.cloud.pubsub.v1.{TopicAdminClient, TopicAdminSettings}
@@ -96,9 +95,11 @@ object GoogleTopicAdminInterpreter {
     credential: ServiceAccountCredentials
   ): Resource[F, TopicAdminClient] = {
     val executorProviderBuilder = TopicAdminSettings.defaultExecutorProviderBuilder()
-    val executorProvider = getExecutorProvider(executorProviderBuilder, "goog2-topic-admin-%d")
-    val transportProvider =
-      TopicAdminSettings.defaultTransportChannelProvider().withExecutor(executorProvider.getExecutor)
+    val threadFactory = new ThreadFactoryBuilder()
+      .setThreadFactory(executorProviderBuilder.getThreadFactory)
+      .setNameFormat("goog2-topic-admin-%d")
+      .build()
+    val executorProvider = executorProviderBuilder.setThreadFactory(threadFactory).build()
 
     Resource.make(
       Async[F].delay(
@@ -107,7 +108,6 @@ object GoogleTopicAdminInterpreter {
             .newBuilder()
             .setCredentialsProvider(FixedCredentialsProvider.create(credential))
             .setBackgroundExecutorProvider(executorProvider)
-            .setTransportChannelProvider(transportProvider)
             .build()
         )
       )